# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"TasksResponse": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITasksPayload": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "header"
			}
		}
	},
	"info": {
		"description": "Nodejs+Express+Typescript+TSOA Example",
		"version": "1.0.0",
		"title": "Example Nodejs+Typescript"
	},
	"openapi": "3.0.0",
	"paths": {
		"/tasks": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TasksResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateMessage",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TasksResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITasksPayload"
							}
						}
					}
				}
			}
		},
		"/tasks/{id}": {
			"put": {
				"operationId": "UpdateMessage",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TasksResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITasksPayload"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMessage",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TasksResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:8080",
			"description": "Development server"
		},
		{
			"url": "http://staging:8080",
			"description": "Staging server"
		},
		{
			"url": "http://production:8080",
			"description": "Production server"
		}
	]
}